+++ CHANGES.md
@@ -71,6 +71,9 @@ To be released.
  -  Added `Move` class to Activity Vocabulary API.
     [[#65], [#92] by Lee Dogeon]
 
+ -  Added `Read` class to Activity Vocabulary API.
+    [[#65], [#92] by Lee Dogeon]
+
 [#53]: https://github.com/dahlia/fedify/issues/53
 [#66]: https://github.com/dahlia/fedify/issues/66
 [#70]: https://github.com/dahlia/fedify/issues/70